Changed how CalcUsedWorkingThreadCount function works: - New parameter macroParamCase added to tell if this is macroParam calculation or editing edited data case - If editing edited data, using 75 % of system threads because all the working threads available in long lasting editing are beneficial - If macroParam case, only 3 working thread are used (unless userGivenWorkingThreadCount has non-zero value). MacroParam calculations have only one time step and usage of multiple working threads in that case is usually waste of effort. Especially when a lot of station data parameters are involved, then each thread have to have copies of each stations  (which can be thousands in Synop case) and possible multiple times for each calculation block.
Using 3 working threads is the first best guess for grid data and station data usage. For some cases with station data best results come with 1 working thread, but using 3 is not that bad. In many cases with grid data parameters, about 3 working thread is quite optimal, but never the original of 75 % of e.g. 24 total available thread (18).